Remove OpenAI API key from workflows for security
