+++ .github/workflows/Agents.yml
@@ -17,7 +17,6 @@ jobs:
       DATADOG_API_KEY: ${{ secrets.DATADOG_API_KEY }}
       XMTP_ENV: ${{ matrix.environment }}
       GEOLOCATION: ${{ vars.GEOLOCATION }}
-      OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
       SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
       WALLET_KEY: ${{ secrets.WALLET_KEY }}
       DEFAULT_STREAM_TIMEOUT_MS: 8000
