+++ .github/workflows/Large.yml
@@ -20,7 +20,6 @@ jobs:
       WORKER_COUNT: 5
       BATCH_SIZE: 50
       MAX_GROUP_SIZE: 200
-      OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
       SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
       SLACK_CHANNEL: ${{ secrets.SLACK_CHANNEL }}
     steps:
