+++ helpers/playwright.ts
@@ -7,7 +7,7 @@ import {
   type BrowserContext,
   type Page,
 } from "playwright-chromium";
-import { defaultValues } from "./tests";
+import { defaultValues } from "./utils";
 
 // Default timeout for stream collection in milliseconds
 const DEFAULT_STREAM_TIMEOUT_MS = process.env.DEFAULT_STREAM_TIMEOUT_MS
