+++ suites/functional/codec.test.ts
@@ -1,5 +1,5 @@
 import { loadEnv } from "@helpers/client";
-import { getFixedNames } from "@helpers/tests";
+import { getFixedNames } from "@helpers/utils";
 import { setupTestLifecycle } from "@helpers/vitest";
 import { getWorkers, type WorkerManager } from "@workers/manager";
 import {
