+++ suites/functional/streams.test.ts
@@ -8,7 +8,7 @@ import {
   verifyMetadataStream,
   verifyNewConversationStream,
 } from "@helpers/streams";
-import { getFixedNames, getInboxIds } from "@helpers/tests";
+import { getFixedNames, getInboxIds } from "@helpers/utils";
 import { setupTestLifecycle } from "@helpers/vitest";
 import { typeofStream } from "@workers/main";
 import { getWorkers } from "@workers/manager";
